dev: Update CODEOWNERS and review structure
