+++ pkg/joinserver/joinserver.go
@@ -49,14 +49,6 @@ import (
 	"google.golang.org/grpc"
 )
 
-// Config represents the JoinServer configuration.
-type Config struct {
-	Devices         DeviceRegistry      `name:"-"`
-	Keys            KeyRegistry         `name:"-"`
-	JoinEUIPrefixes []types.EUI64Prefix `name:"join-eui-prefix" description:"JoinEUI prefixes handled by this JS"`
-	DeviceKEKLabel  string              `name:"device-kek-label" description:"Label of KEK used to encrypt device keys at rest"`
-}
-
 // JoinServer implements the Join Server component.
 //
 // The Join Server exposes the NsJs and DeviceRegistry services.
